More compact, efficent solution to storage using update instead of force write.
According to the library's documentation:
"In addition to the write functions there is a set of update ones. This functions read each byte first and skip the burning if the old value is the same with new. The scaning direction is from high address to low, to obtain quick return in common cases."